Evaluator and some modifications

+++ Stylometry/TrainedData.cs
@@ -9,14 +9,14 @@ namespace Stylometry
     public class TrainedData
     {
         public int WordFrequency { get; set; }
-        public int AverageLetterLength { get; set; }
+        public float AverageLetterLength { get; set; }
         public float NounFrequency { get; set; }
         public float VerbFrequency { get; set; }
         public float MostCommonWordCount { get; set; }
         public float TagsDiversity { get; set; }
         public int AuthorId { get; set; }
 
-        public TrainedData(int wordFrequency, int averageWordFrequency,
+        public TrainedData(int wordFrequency, float averageWordFrequency,
             float nounFrequency, float verbFrequency, float mostCommonWordCount, float tagsDiversity, int authorId)
         {
             WordFrequency = wordFrequency;
